+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Remove `output_enable` options, add `duty_offset` option instead
     * Duty ratio is now calculated by (`duty`+`duty_offset`)/512 %
 * Merge mod-from-file
+* Add emulator link (experimental)
 * Update firmware version to 1.3
 
 # 1.4.3